ModernMedical
If the number is divisible by 3 only, output Modern
c. If the number is divisible by 5 only, output Medical
d. If the number is divisible by both 3 and 5, output ModernMedical
e. If the number is not divisible by either 3 or 5, only output the
number itself.
f. The correct output for the first 20 of the 100 numbers should look
like:

1
2
Modern
4
Medical
Modern
7
8
Modern
Medical
11
Modern
13
14
ModernMedical
16
17
Modern
19
Medical